requirements.txt
google-cloud-bigquery[pandas]==3.26.0
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
plotly==5.24.1
scikit-learn==1.5.2
pyarrow==18.0.0
dash==2.18.2